Yatto solved...

+++ abc175_c.jl
@@ -21,7 +21,7 @@ end
 
 parseint() = readline() |> x -> parse(Int, x)
 parsestring() = readline()
-parseints() = readline() |> split |> x -> parse.(Int, x)
+parseints() = readline() |> split |> x -> parse.(BigInt, x)
 parsestrings() = readline() |> split
 
 main()
